move find_package(catalyst) to lower src/CMakeList
In order to not require the GLOBAL keyword and keep the imported targets
local find_package(catalyst) is moved to the lower list.
Then catalyst::catalyst is available and can be used for linkage.

Using cmake-generator-expression [0] to only include compile definition
as well as linkage if the ENABLE_CATALYST flag is set, otherwise the
return an empy string.
This prevents from tidious IF .. ELSEIF .. ENDIF construct for linkage.

[0] : https://cmake.org/cmake/help/latest/manual/cmake-generator-expressions.7.html