Added function to compute fem RHS vector from particle distribution. The added function computes the value of the basis function for each particle and adds this value to the entry of that basis function.
The unit tests for this create a mesh with random origin and spacing and sample a fixed number of particles uniformly, project those and compute if the total sum over the dofs is the same as the sum of the particle values. (This can be seen as charge-conservation). This type of test only works if the sum of the basis functions are a partition of unity.

In LagrangeSpace.hpp, the function getGlobalDOFIndices was modified and the irrelevant mapping was removed.